Revert "Feat/register"
