+++ src/components/Router.js
@@ -3,6 +3,7 @@ import {
   HashRouter as Router,
   Routes,
   Route,
+  Outlet,
   Link,
   Navigate,
 } from "react-router-dom";
@@ -11,14 +12,10 @@ import Home from "routes/Home";
 import Profile from "routes/Profile";
 import Navigation from "components/Navigation";
 
-export default function AppRouter({
-  isLoggedIn,
-  userObj,
-  refreshUser,
-  codeNum,
-}) {
+export default function AppRouter({ isLoggedIn, userObj, refreshUser }) {
   return (
     <div>
+      <h1>사외공사장 테스트</h1>
       {/* Routes nest inside one another. Nested route paths build upon
             parent route paths, and nested route elements render inside
             parent route elements. See the note about <Outlet> below. */}
@@ -27,10 +24,7 @@ export default function AppRouter({
           {isLoggedIn ? (
             <>
               <Route path="/" element={<Navigation />}>
-                <Route
-                  index
-                  element={<Home userObj={userObj} codeNum={codeNum} />}
-                />
+                <Route index element={<Home userObj={userObj} />} />
                 <Route
                   path="profile"
                   element={
