--- src/styles.scss
@@ -1,158 +1,0 @@
-* {
-  box-sizing: border-box;
-}
-
-body {
-  background-color: rgb(248, 248, 248);
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen,
-    Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
-  font-size: 14px;
-  color: black;
-  height: 100vh;
-}
-
-.container {
-  width: 100%;
-  max-width: 95%;
-  display: flex;
-  flex-direction: column;
-}
-.hero {
-  margin-bottom: 20px;
-}
-
-.gongsa__actions {
-  position: absolute;
-  right: 10px;
-  top: 10px;
-}
-
-.boxForm {
-  display: flex;
-  flex-direction: column;
-  margin-top: 20px;
-  margin-bottom: 20px;
-}
-
-.imgLabel {
-  text-align: center;
-}
-
-.box {
-  border-radius: 20px;
-}
-
-.reg-temp-img {
-  right: -10px;
-  top: 20px;
-  border-radius: 50%;
-  width: 70px;
-  height: 70px;
-  margin-top: 10px;
-}
-
-.edit-gongsa {
-  display: flex;
-  flex-direction: column;
-}
-
-.org-img {
-  right: -10px;
-  top: 20px;
-  border-radius: 5%;
-  margin-top: 10px;
-}
-
-.hr-main {
-  margin-top: 5px;
-}
-
-.box {
-  box-shadow: 0em;
-}
-.box-header {
-  display: flex;
-  flex-direction: row;
-}
-
-.box-header-text {
-  font-size: 19px;
-  font-weight: bold;
-}
-
-.nav-header {
-  display: block;
-}
-
-.nav-content {
-  display: flex;
-  height: 100%;
-}
-
-.nav-brand {
-  flex: 1;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.nav-ul {
-  flex: 1;
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-}
-
-.title {
-  font-size: 24px;
-}
-
-.subtitle {
-  font-size: 12px;
-}
-
-.edit-box {
-  background-color: rgb(248, 248, 248);
-}
-
-.auth-container {
-  display: flex;
-  flex-direction: column;
-  height: 100vh;
-  justify-content: center;
-  align-items: center;
-}
-
-.img-hide {
-  opacity: 0;
-}
-
-.footer {
-  position: fixed;
-  left: 0;
-  bottom: 0;
-  width: 100%;
-  background-color: rgb(230, 230, 230);
-  text-align: center;
-  padding: 1rem 1rem 1rem;
-}
-
-.reg-btn {
-  margin-top: 20px;
-}
-
-.div-arrow {
-  display: flex;
-  flex: 1;
-  justify-content: flex-end;
-}
-
-.select-region {
-  display: flex;
-  justify-content: flex-end;
-}
-
-.content-gongsa {
-  display: flex;
-  flex-direction: row;
-}
